requirements.txt
fastapi==0.115.0uvicorn[standard]==0.34.0streamlit==1.41.0requests==2.32.3sqlalchemy==2.0.36python-dotenv==1.0.1openai==1.59.6pymupdf==1.24.14passlib==1.7.4bcrypt==4.0.1python-multipart==0.0.9pydantic==2.10.4